+++ jeap-messaging-infrastructure/src/main/java/ch/admin/bit/jeap/messaging/kafka/contract/ContractValidationExemptions.java
@@ -10,16 +10,10 @@ import java.util.Set;
 public class ContractValidationExemptions {
 
     private static final Collection<String> SHARED_MESSAGES_ALLOWED_TO_SEND =
-            Set.of("MessageProcessingFailedEvent", "ProcessTaskPlannedEvent", "ProcessTaskCompletedEvent", "ProcessInstanceCompletedEvent", "CreateProcessInstanceCommand");
-    private static final Collection<String> SHARED_MESSAGES_ALLOWED_TO_RECEIVE =
-            Set.of("ProcessInstanceCreatedEvent", "ProcessMilestoneReachedEvent", "ProcessInstanceCompletedEvent", "CreateProcessInstanceCommand");
+            Set.of("MessageProcessingFailedEvent", "ReactionIdentifiedEvent", "ReactionsObservedEvent");
 
     public static boolean isExemptedFromSenderValidation(MessageType messageType) {
         return SHARED_MESSAGES_ALLOWED_TO_SEND.contains(messageType.getName());
-    }
-
-    public static boolean isExemptedFromReceiverValidation(String messageTypeName) {
-        return SHARED_MESSAGES_ALLOWED_TO_RECEIVE.contains(messageTypeName);
     }
 
 }