+++ CHANGELOG.md
@@ -5,6 +5,13 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/), and this project adheres
 to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [8.32.0] - 2025-04-23
+
+### Changed
+
+- Add contract exemptions for ReactionIdentifiedEvent and ReactionsObservedEvent
+- Remove deprecated / unnecessary contract exemptions for process context service events
+
 ## [8.31.2] - 2025-04-03
 
 ### Changed